Core:  * Added Testsuite to branch => Many bugfixes  * Changed reporting interface
CMSSW:
 * Option to choose instrumentation fragment



git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/grid-control/trunk/grid-control@926 29098800-e21f-4f28-a8ff-84ee76f0ac90

+++ python/grid_control/backends/grid_wms.py
@@ -1,12 +1,17 @@
 from python_compat import *
-import sys, os, time, copy, tempfile, cStringIO, tarfile
-from grid_control import ConfigError, RethrowError, Job, utils
+import os, time, copy, tempfile, tarfile
+from grid_control import ConfigError, APIError, RethrowError, Job, utils
 from wms import WMS
 
 try:
 	from email.utils import parsedate
 except ImportError:
 	from email.Utils import parsedate
+
+def jdlEscape(value):
+	repl = { '\\': r'\\', '\"': r'\"', '\n': r'\n' }
+	return '"' + str.join('', map(lambda char: repl.get(char, char), value)) + '"'
+
 
 class GridWMS(WMS):
 	_statusMap = {
@@ -25,80 +30,55 @@ class GridWMS(WMS):
 
 
 	def __init__(self, config, module, monitor, section):
-		WMS.__init__(self, config, module, monitor, 'grid', 'VomsProxy')
+		config.set('grid', 'proxy', 'VomsProxy', override = False)
+		WMS.__init__(self, config, module, monitor, 'grid', None)
 		self._sites = config.get('grid', 'sites', '', volatile=True).split()
 		self.vo = config.get('grid', 'vo', self.proxy.getVO())
 
 		self._submitParams = {}
 		self._ce = config.get(section, 'ce', '', volatile=True)
 		self._configVO = config.getPath(section, 'config', '', volatile=True)
-		if self._configVO != '' and not os.path.exists(self._configVO):
-			raise ConfigError("--config file '%s' does not exist." % self._configVO)
-
-
-	def _jdlEscape(value):
-		repl = { '\\': r'\\', '\"': r'\"', '\n': r'\n' }
-		def replace(char):
-			try:
-				return repl[char]
-			except:
-				return char
-		return '"' + str.join('', map(replace, value)) + '"'
-	_jdlEscape = staticmethod(_jdlEscape)
 
 
 	def storageReq(self, sites):
-		fmt = lambda x: 'Member(%s, other.GlueCESEBindGroupSEUniqueID)' % self._jdlEscape(x)
+		fmt = lambda x: 'Member(%s, other.GlueCESEBindGroupSEUniqueID)' % jdlEscape(x)
 		if sites:
 			return '( %s )' % str.join(' || ', map(fmt, sites))
 
 
 	def sitesReq(self, sites):
-		sitereqs = []
-		fmt = lambda x: 'RegExp(%s, other.GlueCEUniqueID)' % self._jdlEscape(x)
-
-		blacklist = filter(lambda x: x.startswith('-'), sites)
-		sitereqs.extend(map(lambda x: '!' + fmt(x[1:]), blacklist))
-
-		whitelist = filter(lambda x: not x.startswith('-'), sites)
+		fmt = lambda x: 'RegExp(%s, other.GlueCEUniqueID)' % jdlEscape(x)
+		(blacklist, whitelist) = utils.splitBlackWhiteList(sites[1])
+		sitereqs = map(lambda x: '!' + fmt(x), blacklist)
 		if len(whitelist):
 			sitereqs.append('(%s)' % str.join(' || ', map(fmt, whitelist)))
-
 		if sitereqs:
 			return '( %s )' % str.join(' && ', sitereqs)
 
 
 	def _formatRequirements(self, reqs):
 		result = ['other.GlueHostNetworkAdapterOutboundIP']
-		for type, arg in reqs:
-			if type == self.SOFTWARE:
-				result.append('Member(%s, other.GlueHostApplicationSoftwareRunTimeEnvironment)' % self._jdlEscape(arg))
-			elif type == self.WALLTIME:
+		for reqType, arg in reqs:
+			if reqType == WMS.SOFTWARE:
+				result.append('Member(%s, other.GlueHostApplicationSoftwareRunTimeEnvironment)' % jdlEscape(arg))
+			elif reqType == WMS.WALLTIME:
 				if arg > 0:
 					result.append('(other.GlueCEPolicyMaxWallClockTime >= %d)' % int((arg + 59) / 60))
-			elif type == self.CPUTIME:
+			elif reqType == WMS.CPUTIME:
 				if arg > 0:
 					result.append('(other.GlueCEPolicyMaxCPUTime >= %d)' % int((arg + 59) / 60))
-			elif type == self.MEMORY:
+			elif reqType == WMS.MEMORY:
 				if arg > 0:
 					result.append('(other.GlueHostMainMemoryRAMSize >= %d)' % arg)
-			elif type == self.STORAGE:
+			elif reqType == WMS.STORAGE:
 				result.append(self.storageReq(arg))
-			elif type == self.SITES:
+			elif reqType == WMS.SITES:
 				result.append(self.sitesReq(arg))
-			elif type == self.CPUS:
+			elif reqType == WMS.CPUS:
 				pass # Handle number of cpus in makeJDL
 			else:
-				raise RuntimeError('unknown requirement type %s or argument %r' % (WMS.reqTypes[type], arg))
+				raise APIError('Unknown requirement type %s or argument %r' % (WMS.reqTypes[reqType], arg))
 		return str.join(' && ', filter(lambda x: x != None, result))
-
-
-	def getRequirements(self, jobNum):
-		reqs = WMS.getRequirements(self, jobNum)
-		# add site requirements
-		if len(self._sites):
-			reqs.append((self.SITES, self._sites))
-		return reqs
 
 
 	def makeJDL(self, jobNum):
@@ -111,10 +91,8 @@ class GridWMS(WMS):
 			self.writeJobConfig(jobNum, cfgPath)
 			sandboxOutJDL = self.sandboxOut
 
-		def formatStrList(strList):
-			return '{ ' + str.join(', ', map(lambda x: '"%s"' % x, strList)) + ' }'
-
-		reqs = self.getRequirements(jobNum)
+		reqs = self.broker.brokerSites(self.module.getRequirements(jobNum))
+		formatStrList = lambda strList: '{ ' + str.join(', ', map(lambda x: '"%s"' % x, strList)) + ' }'
 		contents = {
 			'Executable': '"gc-run.sh"',
 			'Arguments': '"%d"' % jobNum,
@@ -127,50 +105,10 @@ class GridWMS(WMS):
 			'Rank': '-other.GlueCEStateEstimatedResponseTime',
 			'RetryCount': 2
 		}
-		cpus = dict(reqs).get(self.CPUS, 1)
+		cpus = dict(reqs).get(WMS.CPUS, 1)
 		if cpus > 1:
 			contents['CpuNumber'] = cpus
 		return utils.DictFormat(' = ').format(contents, format = '%s%s%s;\n')
-
-
-	def cleanup(self, list):
-		for item in list:
-			try:
-				if os.path.isdir(item):
-					os.rmdir(item)
-				else:
-					os.unlink(item)
-			except:
-				pass
-
-
-	def logError(self, proc, **kwargs):
-		retCode, stdout, stderr = proc.getAll()
-		kwargs.update({'stdout': stdout, 'stderr': stderr})
-		utils.eprint('WARNING: %s failed with code %d' % (os.path.basename(proc.cmd[0]), retCode))
-
-		now = time.time()
-		entry = '%s.%s' % (time.strftime('%Y-%m-%d_%H:%M:%S', time.localtime(now)), ('%.5f' % (now - int(now)))[2:])
-		data = { 'retCode': retCode, 'exec': proc.cmd[0], 'args': proc.cmd[1] }
-		sys.stderr.writelines(filter(lambda x: (x != '\n') and not x.startswith('----'), stderr))
-
-		tar = tarfile.TarFile.open(os.path.join(self.config.workDir, 'error.tar'), 'a')
-		
-		files = [utils.VirtualFile(os.path.join(entry, 'info'), utils.DictFormat().format(data))]
-		for name, path in kwargs.items():
-			try:
-				content = open(path, 'r').readlines()
-			except:
-				content = []
-			files.append(utils.VirtualFile(os.path.join(entry, name), content))
-		for file in files:
-			info, handle = file.getTarInfo()
-			tar.addfile(info, handle)
-			handle.close()
-			sys.stderr.write('.')
-		tar.close()
-		sys.stderr.write('\nAll logfiles were moved to %s\n' % os.path.join(self.config.workDir, 'error.tar'))
-		return False
 
 
 	def writeWMSIds(self, ids):
@@ -241,16 +179,14 @@ class GridWMS(WMS):
 
 
 	def _parseStatusX(self, lines):
-		buffer = []
-		for line in lines:
-			bline = line.strip('*\n')
-			if bline != '' and ('BOOKKEEPING INFORMATION' not in bline):
-				buffer.append(bline)
-			if line.startswith('****') and len(buffer):
-				remap = { 'destination': 'dest', 'status reason': 'reason',
-					'status info for the job': 'id', 'current status': 'status',
-					'submitted': 'timestamp', 'reached': 'timestamp', 'exit code': 'gridexit'  }
-				data = utils.DictFormat(':').parse(buffer, keyRemap = remap)
+		adder = lambda a,b: utils.QM('=====' not in b and b != '\n', a+b, a)
+		remap = { 'destination': 'dest', 'status reason': 'reason',
+			'status info for the job': 'id', 'current status': 'status',
+			'submitted': 'timestamp', 'reached': 'timestamp', 'exit code': 'gridexit'  }
+		for section in utils.accumulate(lines, lambda x, buf: ('='*70) in x, opAdd = adder):
+			data = utils.DictFormat(':').parse(str.join('', section), keyRemap = remap)
+			data = utils.filterDict(data, vF = lambda v: v)
+			if data:
 				try:
 					if 'failed' in data['status']:
 						data['status'] = 'failed'
@@ -263,7 +199,6 @@ class GridWMS(WMS):
 				except:
 					pass
 				yield data
-				buffer = []
 
 
 	def explainError(self, proc, code):
@@ -275,25 +210,24 @@ class GridWMS(WMS):
 	# Submit job and yield (jobNum, WMS ID, other data)
 	def submitJob(self, jobNum):
 		fd, jdl = tempfile.mkstemp('.jdl')
-		log = tempfile.mktemp('.log')
-
 		try:
 			data = self.makeJDL(jobNum)
 			utils.safeWrite(os.fdopen(fd, 'w'), data)
 		except:
+			utils.removeFiles([jdl])
 			raise RethrowError('Could not write jdl data to %s.' % jdl)
 
-		tmp = utils.filterDict(self._submitParams, vF = lambda v: v != '')
-		params = str.join(' ', map(lambda (x, y): '%s %s' % (x, y), tmp.items()))
-
 		try:
+			tmp = utils.filterDict(self._submitParams, vF = lambda v: v != '')
+			params = str.join(' ', map(lambda (x, y): '%s %s' % (x, y), tmp.items()))
+
+			log = tempfile.mktemp('.log')
 			activity = utils.ActivityLog('submitting jobs')
 			proc = utils.LoggedProcess(self._submitExec, '%s --nomsg --noint --logfile "%s" "%s"' % (params, log, jdl))
 
 			wmsId = None
-			for line in map(str.strip, proc.iter()):
-				if line.startswith('http'):
-					wmsId = line
+			for line in filter(lambda x: x.startswith('http'), map(str.strip, proc.iter())):
+				wmsId = line
 			retCode = proc.wait()
 			del activity
 
@@ -301,9 +235,9 @@ class GridWMS(WMS):
 				if self.explainError(proc, retCode):
 					pass
 				else:
-					self.logError(proc, log = log, jdl = jdl)
+					proc.logError(self.errorLog, log = log, jdl = jdl)
 		finally:
-			self.cleanup([log, jdl])
+			utils.removeFiles([log, jdl])
 		return (jobNum, wmsId, {'jdl': str.join('', data)})
 
 
@@ -318,10 +252,8 @@ class GridWMS(WMS):
 
 		activity = utils.ActivityLog('checking job status')
 		proc = utils.LoggedProcess(self._statusExec, '--verbosity 1 --noint --logfile "%s" -i "%s"' % (log, jobs))
-
 		for data in self._parseStatus(proc.iter()):
-			yield (jobNumMap[data['id']], data['id'], self._statusMap[data['status']], data)
-
+			yield (jobNumMap.get(data['id']), data['id'], self._statusMap[data['status']], data)
 		retCode = proc.wait()
 		del activity
 
@@ -329,8 +261,8 @@ class GridWMS(WMS):
 			if self.explainError(proc, retCode):
 				pass
 			else:
-				self.logError(proc, log = log, jobs = jobs)
-		self.cleanup([log, jobs])
+				proc.logError(self.errorLog, log = log, jobs = jobs)
+		utils.removeFiles([log, jobs])
 
 
 	# Get output of jobs and yield output dirs
@@ -348,7 +280,7 @@ class GridWMS(WMS):
 			if not os.path.exists(tmpPath):
 				os.makedirs(tmpPath)
 		except:
-			raise RuntimeError("Temporary path '%s' could not be created." % tmpPath)
+			raise RuntimeError('Temporary path "%s" could not be created.' % tmpPath)
 
 		jobNumMap = dict(ids)
 		jobs = self.writeWMSIds(ids)
@@ -365,14 +297,15 @@ class GridWMS(WMS):
 			if line.startswith(tmpPath):
 				todo.remove(currentJobNum)
 				outputDir = line.strip()
-				if 'GC_WC.tar.gz' in os.listdir(outputDir):
-					wildcardTar = os.path.join(outputDir, 'GC_WC.tar.gz')
-					try:
-						tarfile.TarFile.open(wildcardTar, 'r:gz').extractall(outputDir)
-						os.unlink(wildcardTar)
-					except:
-						utils.eprint("Can't unpack output files contained in %s" % wildcardTar)
-						pass
+				if os.path.exists(outputDir):
+					if 'GC_WC.tar.gz' in os.listdir(outputDir):
+						wildcardTar = os.path.join(outputDir, 'GC_WC.tar.gz')
+						try:
+							tarfile.TarFile.open(wildcardTar, 'r:gz').extractall(outputDir)
+							os.unlink(wildcardTar)
+						except:
+							utils.eprint("Can't unpack output files contained in %s" % wildcardTar)
+							pass
 				yield (currentJobNum, line.strip())
 				currentJobNum = None
 			else:
@@ -382,10 +315,10 @@ class GridWMS(WMS):
 
 		if retCode != 0:
 			if 'Keyboard interrupt raised by user' in proc.getError():
-				self.cleanup([log, jobs, basePath])
+				utils.removeFiles([log, jobs, basePath])
 				raise StopIteration
 			else:
-				self.logError(proc, log = log)
+				proc.logError(self.errorLog, log = log)
 			utils.eprint('Trying to recover from error ...')
 			for dir in os.listdir(basePath):
 				yield (None, os.path.join(basePath, dir))
@@ -394,7 +327,7 @@ class GridWMS(WMS):
 		for jobNum in todo:
 			yield (jobNum, None)
 
-		self.cleanup([log, jobs, basePath])
+		utils.removeFiles([log, jobs, basePath])
 
 
 	def cancelJobs(self, allIds):
@@ -420,11 +353,11 @@ class GridWMS(WMS):
 			# select cancelled jobs
 			for deletedWMSId in filter(lambda x: x.startswith('- '), proc.iter()):
 				deletedWMSId = deletedWMSId.strip('- \n')
-				yield (deletedWMSId, jobNumMap[deletedWMSId])
+				yield (deletedWMSId, jobNumMap.get(deletedWMSId))
 
 			if retCode != 0:
 				if self.explainError(proc, retCode):
 					pass
 				else:
-					self.logError(proc, log = log)
-			self.cleanup([log, jobs])
+					proc.logError(self.errorLog, log = log)
+			utils.removeFiles([log, jobs])
